Move preplanning saves into BLT saves folder
This was a bit difficult, because we still put all of the saves into a
preplanning folder inside the saves directory. To do this, I had to call
os.execute(). For some reason, despite the fact that it succeeds, it
causes my game to minimise.

I also tidied up some of the code. The success messages will now only
appear when things actually succeeded. It won't attempt to load
preplanning data if there is no data available.