Add support for saving assets. Refactor most saving code into functions.
The credit for saving assets goes to r3ddr4gOn on Github. Additionally,
the ideas for checking to see if a table is empty before trying to
iterate over it, as well as making sure chat is available before trying
to use it, are also both ideas that were originally r3ddr4gOn. I have
adapted them to this fork of Preplanned.